Unify error handling for different error responses
We are trying to detect the type of error response and return back the
valid error based on the type. This should be unified in API though so
we have one common error response.

Minor changes for the error output as we print two things and it should
be unified across all commands in next stage:
- golog -> no need to have this in the output
- formatted and informative error as cobra prints the errors
automatically for us.

Example for the error to return
```
indexes, err := client.GetIndexes(connectionName, includeSystemIndexes)
if err != nil {
	return fmt.Errorf("Failed to retrieve indexes. Error: [%s]",
err.Error())
}
```