feat/DATA-1078 Expect tags to be wrapped into a struct
As per agreed API contract, tags should not be exposed as plain
strings but they should be wrapped into structs. This should help in
keeping the API forward compatible as we can easily add new fields to
the JSON representation of the tag.

see https://github.com/lensesio-dev/documentation-gen1/pull/37