Refactor 'import topics'
Currently when user tries to run `import topics` we do a simple check by topic
name, if it exits in Lenses instance then we perform a `PUT` operation if not
a `POST`. Regardless if there are changes between the yaml definition of a topic
and the Lenses topic object we perform this operation leading to excessive API
usage and event logging. This new implementation make a comparison of
the topics partitions and config and performs a PUT only when a diff is
found.

Issue: https://landoop.atlassian.net/browse/OPS-1839