Fix create/update policy bug (flags vs. file)
When creating or updating a policy you can either pass the values as
flags or as a file. In the case of flags, for "fields" and "datasets" we
need to map their string value to slice of strings when ',' has been
used a delimiter. It was found that this was not consistent between the
create and update command causing paylod disrepancies.

Signed-off-by: Konstantinos 'Dean' Linaras <linaras@pm.me>