Introduce tailored connection management commands
Currently, the connection command is generic. A user is supposed to
provide a JSON blob for a particular connection type and the CLI
will pass it through to the API. The format of the JSON blob is
to be obtained from the API docs.

This commit introduces for every connection type a specific set
of commands to manage connections of that type. Attributes of the
connection are not provided via JSON but rather via flags. The
"requiredness" and types (string, string list, bool, etc.) of the
flags corresponds to that of the attributes.

This makes this part of the CLI less prone to errors and
self-documenting.